+++ packages/patterns/NEWS.md
@@ -1,5 +1,10 @@
 User-visible changes in `@endo/patterns`:
 
+# next release
+
+- Add `M.tagged(tagPattern, payloadPattern)` for making patterns that match
+  Passable Tagged objects.
+
 # v0.2.6 (2023-09-11)
 
 - Adds support for CopyMap patterns (e.g., `matches(specimen, makeCopyMap([]))`).